+++ CHANGES.md
@@ -1,2 +1,8 @@
+#### Version 2
+
+- Breaking change: `symbol.offset` renamed to `symbol.address`
+- Added `GET /v1/functions` endpoint
+
 #### Version 1
+
 - Initial release